start decomp cli; adapters->settings; fix shop-prices, ctoptions
Start decomposing arguments.py under sourcefiles/cli. Add missing
ctoptions and fix --shop-prices for CLI, and cover in unit tests.
Fix off-by-one error with battle_speed.
Implement Adapter classes for converting CLI args to Settings.
