+++ sourcefiles/arguments.py
@@ -1,226 +1,18 @@
 from __future__ import annotations
 import argparse
 import copy
-import functools
-import operator
 
-from dataclasses import dataclass
 from pathlib import Path
-from typing import Any, Dict, Iterable, Optional, Union
+from typing import Iterable, Optional, Union
 
+import cli.adapters as adp
 import ctstrings
 import ctoptions
 import randosettings as rset
-from randosettings import GameFlags as GF, GameMode as GM, \
-    CosmeticFlags as CF
 
-SettingsFlags = Union[rset.GameFlags, rset.CosmeticFlags]
+from cli.constants import FLAG_ENTRY_DICT, MYSTERY_FLAG_PROB_ENTRIES
+from randosettings import GameFlags as GF, GameMode as GM, CosmeticFlags as CF
 
-
-@dataclass
-class FlagEntry:
-    name: str = ""
-    short_name: Optional[str] = None
-    help_text: Optional[str] = None
-
-
-_flag_entry_dict: Dict[SettingsFlags, FlagEntry] = {
-    GF.FIX_GLITCH: FlagEntry(
-        "--fix-glitch", "-g",
-        "disable save anywhere and HP overflow glitches"),
-    GF.BOSS_SCALE: FlagEntry(
-        "--boss-scale", "-b",
-        "scale bosses based on key-item locations"),
-    GF.ZEAL_END: FlagEntry(
-        "--zeal-end", "-z",
-        "allow the game to be won when Zeal is defeated in the "
-        "Black Omen"),
-    GF.FAST_PENDANT: FlagEntry(
-        "--fast-pendant", "-p",
-        "the pendant will be charged when 2300 is reached"),
-    GF.LOCKED_CHARS: FlagEntry(
-        "--locked-chars", "-c",
-        "require dreamstone for the dactyl character and factory for "
-        "the Proto Dome character"),
-    GF.UNLOCKED_MAGIC: FlagEntry(
-        "--unlocked-magic", "-m",
-        "magic is unlocked from the beginning of the game without "
-        "visiting Spekkio"),
-    GF.CHRONOSANITY: FlagEntry(
-        "--chronosanity", "-cr",
-        "key items may be found in treasure chests"),
-    GF.ROCKSANITY: FlagEntry(
-        "--rocksanity", None,
-        "rocks are added as key items and key items may be found "
-        "in rock locations"),
-    GF.TAB_TREASURES: FlagEntry(
-        "--tab-treasures", None,
-        "all treasure chests contain tabs"),
-    GF.BOSS_RANDO: FlagEntry(
-        "--boss-randomization", "-ro",
-        "randomize the location of bosses and scale based on location"),
-    GF.CHAR_RANDO: FlagEntry(
-        "--char-rando", "-rc",
-        "randomize character identities and models"), 
-    GF.DUPLICATE_CHARS: FlagEntry(
-        "--duplicate-characters", "-dc",
-        "allow multiple copies of a character to be present in a seed"),
-    GF.DUPLICATE_TECHS: FlagEntry(
-        "--duplicate-techs", None,
-        "allow duplicate characters to perform dual techs together"),
-    GF.VISIBLE_HEALTH: FlagEntry(
-        "--visible-health", None,
-        "the sightscope effect will always be present"),
-    GF.FAST_TABS: FlagEntry(
-        "--fast-tabs", None,
-        "picking up a tab will not pause movement for the fanfare"),
-    GF.BUCKET_LIST: FlagEntry(
-        "--bucket-list", "-k",
-        "allow the End of Time bucket to Lavos to activate when enough "
-        "objectives have been completed."),
-    GF.TECH_DAMAGE_RANDO: FlagEntry(
-        "--tech-damage-rando", None,
-        "Randomize the damage dealt by single techs."),
-    GF.MYSTERY: FlagEntry(
-        "--mystery", None,
-        "choose flags randomly according to mystery settings"),
-    GF.BOSS_SIGHTSCOPE: FlagEntry(
-        "--boss-sightscope", None,
-        "allow the sightscope to work on bosses"),
-    GF.USE_ANTILIFE: FlagEntry(
-        "--use-antilife", None,
-        "use Anti-Life instead of Black Hole for Magus"),
-    GF.TACKLE_EFFECTS_ON: FlagEntry(
-        "--tackle-on-hit-effects", None,
-        "allow Robo Tackle to use the on-hit effects of Robo's weapons"),
-    GF.HEALING_ITEM_RANDO: FlagEntry(
-        "--healing-item-rando", "-he",
-        "randomizes effects of healing items"),
-    GF.FREE_MENU_GLITCH: FlagEntry(
-        "--free-menu-glitch", None,
-        "provides a longer window to enter the menu prior to Lavos3 and "
-        "Zeal2"),
-    GF.GEAR_RANDO: FlagEntry(
-        "--gear-rando", "-q",
-        "randomizes effects on weapons, armors, and accessories"),
-    GF.STARTERS_SUFFICIENT: FlagEntry(
-        "--starters-sufficient", None,
-        "go mode will be acheivable without recruiting additional "
-        "characters"),
-    GF.EPOCH_FAIL: FlagEntry(
-        "--epoch-fail", "-ef",
-        "Epoch flight must be unlocked by bringing the JetsOfTime to "
-        "Dalton in the Snail Stop"),
-    GF.BOSS_SPOT_HP: FlagEntry(
-        "--boss-spot-hp",
-        "boss HP is set to match the vanilla boss HP in each spot"),
-    # Logic Tweak flags from VanillaRando mode
-    GF.UNLOCKED_SKYGATES: FlagEntry(
-        "--unlocked-skyways", None,
-        "Skyways are available as soon as 12kBC is. Normal go mode is still "
-        "needed to unlock the Ocean Palace."),
-    GF.ADD_SUNKEEP_SPOT: FlagEntry(
-        "--add-sunkeep-spot", None,
-        "Adds Sun Stone as an independent key item.  Moonstone charges to a "
-        "random item"),
-    GF.ADD_BEKKLER_SPOT: FlagEntry(
-        "--add-bekkler-spot", None,
-        "C.Trigger unlocks clone game for a KI"),
-    GF.ADD_CYRUS_SPOT: FlagEntry(
-        "--add-cyrus-spot", None,
-        "Gain a KI from Cyrus's Grave w/ Frog.  No Frog stat boost."),
-    GF.RESTORE_TOOLS: FlagEntry(
-        "--restore-tools", None,
-        "Adds Tools. Tools will fix Norther Ruins."),
-    GF.ADD_OZZIE_SPOT: FlagEntry(
-        "--add-ozzie-spot", None, "Gain a KI after Ozzie's Fort."),
-    GF.RESTORE_JOHNNY_RACE: FlagEntry(
-        "--restore-johnny-race", None,
-        "Add bike key and Johnny Race. Bike Key is required to cross Lab32."),
-    GF.ADD_RACELOG_SPOT: FlagEntry(
-        "--add-racelog-spot", None,
-        "Gain a KI from the vanilla Race Log chest."),
-    GF.REMOVE_BLACK_OMEN_SPOT: FlagEntry(
-        "--remove-black-omen-spot", None,
-        "Removes Black Omen rock chest being a possible KI."),
-    GF.SPLIT_ARRIS_DOME: FlagEntry(
-        "--split-arris-dome", None,
-        "Get one key item from the dead guy after Guardian.  Get a second "
-        "after checking the Arris dome computer and bringing the Seed "
-        "(new KI) to Doan."),
-    GF.VANILLA_ROBO_RIBBON: FlagEntry(
-        "--vanilla-robo-ribbon", None,
-        "Gain Robo stat boost from defeating AtroposXR.  If no Atropos in "
-        "seed, then gain from Geno Dome."),
-    GF.VANILLA_DESERT: FlagEntry(
-        "--vanilla-desert", None,
-        "The sunken desert only unlocks after talking to the plant lady "
-        "in Zeal"),
-    # Cosmetic Flags
-    CF.AUTORUN: FlagEntry(
-        "--autorun", None,
-        "Automatically run.  Push run button to walk."
-    ),
-    CF.DEATH_PEAK_ALT_MUSIC: FlagEntry(
-        "--death-peak-alt-music", None,
-        "use Singing Mountain track on Death Peak"
-    ),
-    CF.ZENAN_ALT_MUSIC: FlagEntry(
-        "--zenan-alt-music", None,
-        "use alt battle theme for Zenan Bridge"
-    ),
-    CF.QUIET_MODE: FlagEntry(
-        "--quiet", None,
-        "disable all music (not sound effects)"
-    ),
-    CF.REDUCE_FLASH: FlagEntry(
-        "--reduce-flashes", None,
-        "disable most flashing effects"
-    )
-}
-
-#flag, name, default
-_mystery_flag_prob_entries = [
-    (GF.TAB_TREASURES, "flag_tab_treasures", 0.10),
-    (GF.UNLOCKED_MAGIC, "flag_unlocked_magic", 0.50),
-    (GF.BUCKET_LIST, "flag_bucket_list", 0.15),
-    (GF.CHRONOSANITY, "flag_chronosanity", 0.30),
-    (GF.BOSS_RANDO, "flag_boss_rando", 0.50),
-    (GF.BOSS_SCALE, "flag_boss_scaling", 0.30),
-    (GF.LOCKED_CHARS, "flag_locked_chars", 0.25),
-    (GF.CHAR_RANDO, "flag_char_rando", 0.5),
-    (GF.DUPLICATE_CHARS, "flag_duplicate_chars", 0.25),
-    (GF.EPOCH_FAIL, "flag_epoch_fail", 0.50),
-    (GF.GEAR_RANDO, "flag_gear_rando", 0.25),
-    (GF.HEALING_ITEM_RANDO, "flag_heal_rando", 0.25),
-]
-
-_mode_dict: dict[str, GM] = {
-    'std': GM.STANDARD,
-    'lw': GM.LOST_WORLDS,
-    'loc': GM.LEGACY_OF_CYRUS,
-    'ia': GM.ICE_AGE,
-    'van': GM.VANILLA_RANDO
-}
-
-_diff_dict: dict[str, rset.Difficulty] = {
-    'easy': rset.Difficulty.EASY,
-    'normal': rset.Difficulty.NORMAL,
-    'hard': rset.Difficulty.HARD
-}
-
-_tech_order_dict: dict[str, rset.TechOrder] = {
-    'normal': rset.TechOrder.NORMAL,
-    'balanced': rset.TechOrder.BALANCED_RANDOM,
-    'random': rset.TechOrder.FULL_RANDOM
-}
-
-_shop_price_dict: dict[str, rset.ShopPrices] = {
-    'normal': rset.ShopPrices.NORMAL,
-    'random': rset.ShopPrices.FULLY_RANDOM,
-    'mostrandom': rset.ShopPrices.MOSTLY_RANDOM,
-    'free': rset.ShopPrices.FREE
-}
 
 def add_flags_to_parser(
         group_text: Optional[str],
@@ -236,7 +28,7 @@ def add_flags_to_parser(
         add_target = group
 
     for flag in flag_list:
-        flag_entry = _flag_entry_dict[flag]
+        flag_entry = FLAG_ENTRY_DICT[flag]
         add_args: Iterable[str]
         if flag_entry.short_name is None:
             add_args = (flag_entry.name,)
@@ -247,10 +39,6 @@ def add_flags_to_parser(
             help=flag_entry.help_text,
             action="store_true"
         )
-
-def flag_name_to_namespace_key(flag_name: str):
-    return flag_name[2:].replace('-', '_')
-
 
 # https://stackoverflow.com/questions/3853722/
 # how-to-insert-newlines-on-argparse-help-text
@@ -281,6 +69,21 @@ def get_bucket_settings(args: argparse.Namespace) -> rset.BucketSettings:
         disable_other_go_modes, objectives_win, num_objectives,
         num_objectives_needed, obj_strs
     )
+
+
+def get_ctoptions(args: argparse.Namespace) -> ctoptions.CTOpts:
+    '''Extract CTOpts from argparse.Namespace.'''
+    ct_opts = ctoptions.CTOpts()
+    ct_opts.save_menu_cursor = args.save_menu_cursor
+    ct_opts.save_battle_cursor = args.save_battle_cursor
+    ct_opts.save_tech_cursor = not args.save_skill_cursor_off
+    ct_opts.skill_item_info = not args.skill_item_info_off
+    ct_opts.consistent_paging = args.consistent_paging
+    ct_opts.battle_speed = args.battle_speed - 1
+    ct_opts.battle_msg_speed = args.battle_msg_speed - 1
+    ct_opts.battle_gauge_style = args.battle_gauge_style
+    ct_opts.menu_background = args.background - 1
+    return ct_opts
 
 
 _pc_index_dict: dict[str, int] = {
@@ -357,60 +160,30 @@ def get_mystery_settings(args: argparse.Namespace) -> rset.MysterySettings:
     }
 
     mset.flag_prob_dict = {
-        flag: val_dict['mystery_'+name]
-        for flag, name, _ in _mystery_flag_prob_entries
+        flag: getattr(args, f"mystery_{name}")
+        for flag, name, _ in MYSTERY_FLAG_PROB_ENTRIES
     }
 
     return mset
 
-def fill_flags(val_dict: Dict[str, Any], init: SettingsFlags) -> SettingsFlags:
-    cls = type(init)
-    return functools.reduce(
-        operator.or_,
-        (flag for (flag, entry) in _flag_entry_dict.items()
-         if isinstance(flag, cls)
-         and val_dict[flag_name_to_namespace_key(entry.name)] is True),
-        init
-    )
 
 def args_to_settings(args: argparse.Namespace) -> rset.Settings:
     '''Convert result of argparse to settings object.'''
 
     val_dict = vars(args)
-    flags = rset.GameFlags(fill_flags(val_dict, GF(0)))
-
-    mode = _mode_dict[val_dict['mode']]
-    item_difficulty = _diff_dict[val_dict['item_difficulty']]
-    enemy_difficulty = _diff_dict[val_dict['enemy_difficulty']]
-    tech_order = _tech_order_dict[val_dict['tech_order']]
 
     ret_set = rset.Settings()
-    ret_set.seed = val_dict['seed']
-    ret_set.game_mode = mode
-    ret_set.gameflags = flags
-    ret_set.initial_flags = copy.deepcopy(flags)
-    ret_set.item_difficulty = item_difficulty
-    ret_set.enemy_difficulty = enemy_difficulty
-    ret_set.techorder = tech_order
-
+    ret_set.seed = args.seed
+    ret_set.game_mode = adp.GameModeAdapter.to_setting(args)
+    ret_set.gameflags = adp.GameFlagsAdapter.to_setting(args)
+    ret_set.initial_flags = copy.deepcopy(ret_set.gameflags)
+    ret_set.item_difficulty =  adp.ItemDifficultyAdapter.to_setting(args)
+    ret_set.enemy_difficulty =  adp.EnemyDifficultyAdapter.to_setting(args)
+    ret_set.techorder = adp.TechOrderAdapter.to_setting(args)
+    ret_set.shopprices = adp.ShopPricesAdapter.to_setting(args)
     ret_set.mystery_settings = get_mystery_settings(args)
-    
-    cos_flags = rset.CosmeticFlags(fill_flags(val_dict, CF(0)))
-
-    ct_opts = ctoptions.CTOpts()
-    ct_opts.save_menu_cursor = val_dict['save_menu_cursor']
-    ct_opts.save_battle_cursor = val_dict['save_battle_cursor']
-    ct_opts.save_tech_cursor = not val_dict['save_skill_cursor_off']
-    ct_opts.skill_item_info = not val_dict['skill_item_info_off']
-    ct_opts.consistent_paging = val_dict['consistent_paging']
-    ct_opts.battle_speed = val_dict['battle_speed']
-    ct_opts.battle_msg_speed = val_dict['battle_msg_speed']-1
-    ct_opts.battle_gauge_style = val_dict['battle_gauge_style']
-    ct_opts.menu_background = val_dict['background']-1
-
-    ret_set.cosmetic_flags = cos_flags
-    ret_set.ctoptions = ct_opts
-
+    ret_set.cosmetic_flags = adp.CosmeticFlagsAdapter.to_setting(args)
+    ret_set.ctoptions = get_ctoptions(args)
     ret_set.char_names = [
         val_dict[name.lower()+"_name"] for name in _pc_names
     ]
@@ -458,6 +231,8 @@ def get_parser():
 
     add_generation_options(parser)
 
+    # arguments with a default of "argparse.SUPPRESS", when explicitly specified
+    # on the CLI, will override any other value (e.g. from a preset file)
     parser.add_argument(
         "--mode",
         choices=['std', 'lw', 'ia', 'loc', 'van'],
@@ -468,7 +243,7 @@ def get_parser():
         "  ia: ice age\n"
         " loc: legacy of cyrus\n"
         " van: vanilla rando",
-        default='std',
+        default=argparse.SUPPRESS,
         type=str.lower
     )
 
@@ -477,7 +252,7 @@ def get_parser():
         help="controls quality of treasure, drops, and starting gold "
         "(default: normal)",
         choices=['easy','normal', 'hard'],
-        default='normal',
+        default=argparse.SUPPRESS,
         type=str.lower
     )
 
@@ -486,7 +261,7 @@ def get_parser():
         help="controls strength of enemies and xp/tp rewards "
         "(default: normal)",
         choices=['normal', 'hard'],
-        default='normal',
+        default=argparse.SUPPRESS,
         type=str.lower
     )
 
@@ -498,12 +273,12 @@ def get_parser():
         "balanced - random but biased towards better techs later\n"
         "  random - fully random (default)",
         choices=['normal', 'balanced', 'random'],
-        default='random',
+        default=argparse.SUPPRESS,
         type=str.lower
     )
 
     parser.add_argument(
-        "--shop_prices",
+        "--shop-prices",
         help="R|"
         "controls the prices in shops\n"
         "    normal - standard prices (default)\n"
@@ -511,7 +286,7 @@ def get_parser():
         "mostrandom - random except for staple consumables\n"
         "      free - all items cost 1G",
         choices=['normal', 'random', 'mostrandom', 'free'],
-        default='normal',
+        default=argparse.SUPPRESS,
         type=str.lower
     )
 
@@ -814,7 +589,7 @@ def get_parser():
 
         return fval
 
-    for _, flag_str, prob in _mystery_flag_prob_entries:
+    for _, flag_str, prob in MYSTERY_FLAG_PROB_ENTRIES:
         mystery_flags.add_argument(
             "--mystery_"+flag_str,
             type=check_prob,
